Move `PGRunner` to `test/utils.py` [BF-2080]
To no longer import from `conftest.py` and avoid circular imports.

[BF-2080]
