+++ test/test_db_size_check.py
@@ -1,6 +1,5 @@
 from aiven_db_migrate.migrate.pgmigrate import PGMigrate
-from test.conftest import PGRunner
-from test.utils import random_string
+from test.utils import PGRunner, random_string
 from typing import Tuple
 
 import psycopg2
