+++ test/test_pg_dump.py
@@ -1,8 +1,7 @@
 # Copyright (c) 2020 Aiven, Helsinki, Finland. https://aiven.io/
 
 from aiven_db_migrate.migrate.pgmigrate import PGDatabase, PGMigrate
-from test.conftest import PGRunner
-from test.utils import random_string
+from test.utils import PGRunner, random_string
 from typing import Tuple
 
 import pytest
