+++ test/test_pg_replication.py
@@ -1,8 +1,7 @@
 # Copyright (c) 2020 Aiven, Helsinki, Finland. https://aiven.io/
 
 from aiven_db_migrate.migrate.pgmigrate import PGSource, PGTarget
-from test.conftest import PGRunner
-from test.utils import random_string, Timer
+from test.utils import PGRunner, random_string, Timer
 from typing import Tuple
 
 import psycopg2
