+++ test/test_pg_roles.py
@@ -2,8 +2,7 @@
 from aiven_db_migrate.migrate.errors import PGMigrateValidationFailedError
 from aiven_db_migrate.migrate.pgmigrate import PGMigrate, PGMigrateResult, PGTarget
 from datetime import datetime
-from test.conftest import PGRunner
-from test.utils import modify_pg_security_agent_reserved_roles, random_string
+from test.utils import modify_pg_security_agent_reserved_roles, PGRunner, random_string
 from typing import Tuple
 
 import pytest
