+++ test/test_table_filtering.py
@@ -1,6 +1,5 @@
 from aiven_db_migrate.migrate.pgmigrate import PGMigrate
-from test.conftest import PGRunner
-from test.utils import random_string, Timer
+from test.utils import PGRunner, random_string, Timer
 from typing import Dict, Set, Tuple
 
 import psycopg2
