+++ test/utils.py
@@ -1,15 +1,26 @@
 # Copyright (c) 2020 Aiven, Helsinki, Finland. https://aiven.io/
+from __future__ import annotations
 
 from aiven_db_migrate.migrate.pgmigrate import PGTarget
+from aiven_db_migrate.migrate.pgutils import find_pgbin_dir
 from contextlib import contextmanager
-from test.conftest import PGRunner
-from typing import Iterator
+from datetime import datetime
+from distutils.version import LooseVersion
+from pathlib import Path
+from psycopg2.extras import RealDictCursor
+from typing import Any, Callable, Dict, Iterator, List, Tuple
 
 import datetime
+import logging
+import psycopg2
 import random
+import re
 import string
+import subprocess
 import threading
 import time
+
+SUPPORTED_PG_VERSIONS = ["10", "11", "12", "13", "14"]
 
 
 def random_string(length=20):
@@ -150,6 +161,352 @@ class Timer(TimerBase):
         if not isinstance(self._sleep, tuple):
             return self._sleep
         return random.randrange(*self._sleep)
+
+
+class PGRunner:
+    pgbin: Path
+    pgdata: Path
+    pgport: int
+    pgversion: str
+    aiven_extras_available: bool = False
+    defaultdb: str = "postgres"
+    superuser: str = "postgres"
+    testuser: str = "testuser"
+    cleanups: List[Callable]
+
+    def __init__(self, *, pgversion: str, pgdata: Path, pgport: int = 5432):
+        self.log = logging.getLogger(self.__class__.__name__)
+        self.pgbin = find_pgbin_dir(pgversion)
+        self.pgversion = pgversion
+        self.pgport = pgport
+        self.pgdata = pgdata
+        self.cleanups = list()
+
+    def init(self) -> PGRunner:
+        self.log.info("Initializing pg %s in %r", self.pgversion, self.pgdata)
+        cmd = (
+            self.pgbin / "pg_ctl",
+            "init",
+            "-D",
+            self.pgdata,
+            "-o",
+            "--encoding utf-8",
+        )
+        subprocess.run(cmd, check=True, env={"TZ": "UTC"})
+        return self
+
+    def make_conf(self, **kwargs) -> PGRunner:
+        with open(self.pgdata / "postgresql.conf", "r+") as fp:
+            lines = fp.read().splitlines()
+            fp.seek(0)
+            fp.truncate()
+            config = {}
+            for line in lines:
+                line = line.strip()
+                if not line or line.startswith("#"):
+                    continue
+                key, val = re.split(r"\s*=\s*", line, 1)
+                config[key] = re.sub(r"\s*(#.*)?$", "", val)
+            config.update(
+                # disable fsync and synchronous_commit to speed up the tests a bit
+                fsync="off",
+                synchronous_commit="off",
+                # synchronous_commit="local",
+                # don't need to wait for autovacuum workers when shutting down
+                autovacuum="off",
+                # extensions whitelisting, https://github.com/dimitri/pgextwlist
+                session_preload_libraries="'pgextwlist'",
+            )
+            config.update(kwargs)
+            lines = ["{} = {}\n".format(key, val) for key, val in sorted(config.items())]
+            fp.write("".join(lines))
+        return self
+
+    def make_hba_conf(self, *, dbname: str, user: str, auth: str) -> PGRunner:
+        with open(self.pgdata / "pg_hba.conf", "r+") as fp:
+            lines = fp.read().splitlines()
+            fp.seek(0)
+            fp.truncate()
+            lines = ["{}\n".format(line.strip()) for line in lines if line and not line.startswith("#")]
+            lines.insert(0, f"local\t{dbname}\t{user}\t{auth}\n")
+            fp.write("".join(lines))
+        return self
+
+    def start(self, timeout: int = 10):
+        self.log.info("Starting pg %s in %r", self.pgversion, self.pgdata)
+        cmd: Tuple
+        cmd = (
+            self.pgbin / "pg_ctl",
+            "start",
+            "-D",
+            self.pgdata,
+            "-o",
+            f"-k {self.pgdata} -p {self.pgport} -c listen_addresses=",
+        )
+        subprocess.run(cmd, check=True)
+        # wait until ready to accept connections
+        cmd = (
+            self.pgbin / "pg_isready",
+            "-h",
+            self.pgdata,
+            "-p",
+            str(self.pgport),
+        )
+        timer = Timer(timeout=timeout, what=f"pg {self.pgversion} is ready")
+        while timer.loop(log=self.log):
+            p = subprocess.run(cmd, check=False)
+            if p.returncode == 0:
+                break
+
+    def reload(self):
+        """Re-read configuration files (postgresql.conf, pg_hba.conf, etc.)"""
+        cmd = (
+            self.pgbin / "pg_ctl",
+            "reload",
+            "-D",
+            self.pgdata,
+        )
+        subprocess.run(cmd, check=True)
+
+    def stop(self, timeout: int = 10):
+        self.log.info("Stopping pg %s in %r", self.pgversion, self.pgdata)
+        cmd = (
+            self.pgbin / "pg_ctl",
+            "stop",
+            "-D",
+            self.pgdata,
+            "-m",
+            "smart",
+            "-t",
+            str(timeout),
+        )
+        subprocess.run(cmd, check=True)
+
+    def create_superuser(self, username: str = None):
+        if username is None:
+            username = self.superuser
+        cmd = (
+            self.pgbin / "createuser",
+            "-h",
+            self.pgdata,
+            "-p",
+            str(self.pgport),
+            "--superuser",
+            username,
+        )
+        subprocess.run(cmd, check=True)
+
+    def create_user(self, *, username: str, createdb: bool = True, createrole: bool = True, replication: bool = True):
+        cmd = (
+            self.pgbin / "createuser",
+            "-h",
+            self.pgdata,
+            "-p",
+            str(self.pgport),
+            "--createdb" if createdb else "--no-createdb",
+            "--createrole" if createrole else "--no-createrole",
+            "--replication" if replication else "--no-replication",
+            username,
+        )
+        subprocess.run(cmd, check=True)
+
+    def create_role(
+        self,
+        *,
+        username: str,
+        password: str = None,
+        superuser: bool = False,
+        createdb: bool = True,
+        createrole: bool = True,
+        inherit: bool = True,
+        login: bool = True,
+        replication: bool = False,
+        bypassrls: bool = False,
+        connlimit: int = -1,
+        validuntil: datetime = None,
+        **kwargs,
+    ):
+        if login:
+            assert password, "Password must be set for roles with login"
+        sql = "CREATE ROLE {} WITH {} {} {} {} {} {} {}".format(
+            username,
+            "SUPERUSER" if superuser else "NOSUPERUSER",
+            "CREATEDB" if createdb else "NOCREATEDB",
+            "CREATEROLE" if createrole else "NOCREATEROLE",
+            "INHERIT" if inherit else "NOINHERIT",
+            "LOGIN" if login else "NOLOGIN",
+            "REPLICATION" if replication else "NOREPLICATION",
+            "BYPASSRLS" if bypassrls else "NOBYPASSRLS",
+        )
+        params: List[Any] = []
+        if connlimit != -1:
+            sql += " CONNECTION LIMIT %s"
+            params.append(connlimit)
+        if password:
+            sql += " PASSWORD %s"
+            params.append(password)
+        if validuntil:
+            sql += f" VALID UNTIL '{validuntil}'"
+        config: Dict[str, Any] = {}
+        config.update(kwargs)
+        with self.cursor(username=self.superuser) as cur:
+            cur.execute(sql, params)
+            for key, value in config.items():
+                cur.execute(f"ALTER ROLE {username} SET {key} = %s", (value, ))
+        self.log.info("Created role %r in %r", username, self.pgdata)
+
+    def create_db(self, *, dbname: str, owner: str = None):
+        if owner is None:
+            owner = self.testuser
+        cmd = (
+            self.pgbin / "createdb",
+            "-h",
+            self.pgdata,
+            "-p",
+            str(self.pgport),
+            "--owner",
+            owner,
+            dbname,
+        )
+        subprocess.run(cmd, check=True)
+
+    def conn_info(self, *, username: str = None, dbname: str = None) -> Dict[str, Any]:
+        if username is None:
+            username = self.testuser
+        if dbname is None:
+            dbname = self.defaultdb
+        return {
+            "dbname": dbname,
+            "host": self.pgdata,
+            "port": self.pgport,
+            "user": username,
+        }
+
+    def super_conn_info(self, *, dbname: str = None) -> Dict[str, Any]:
+        return self.conn_info(username=self.superuser, dbname=dbname)
+
+    @contextmanager
+    def cursor(self, *, username: str = None, dbname: str = None, autocommit: bool = True) -> RealDictCursor:
+        conn = None
+        try:
+            conn = psycopg2.connect(**self.conn_info(username=username, dbname=dbname))
+            conn.autocommit = autocommit
+            yield conn.cursor(cursor_factory=RealDictCursor)
+        finally:
+            if conn is not None:
+                conn.close()
+
+    def drop_db(self, *, dbname: str):
+        self.log.info("Dropping database %r from %r", dbname, self.pgdata)
+        cmd = (
+            self.pgbin / "dropdb",
+            "-h",
+            self.pgdata,
+            "-p",
+            str(self.pgport),
+            dbname,
+        )
+        subprocess.run(cmd, check=True)
+
+    def drop_dbs(self):
+        with self.cursor(username=self.superuser) as cur:
+            cur.execute(
+                "SELECT datname from pg_catalog.pg_database WHERE NOT datistemplate AND datname <> %s", (self.defaultdb, )
+            )
+            dbs = cur.fetchall()
+        for db in dbs:
+            self.drop_db(dbname=db["datname"])
+
+    def drop_user(self, *, username):
+        self.log.info("Dropping user %r from %r", username, self.pgdata)
+        cmd = (
+            self.pgbin / "dropuser",
+            "-h",
+            self.pgdata,
+            "-p",
+            str(self.pgport),
+            username,
+        )
+        subprocess.run(cmd, check=True)
+
+    def create_extension(self, *, extname: str, extversion: str = None, dbname: str, grantee: str = None):
+        if grantee is None:
+            grantee = self.testuser
+        sql = f"CREATE EXTENSION IF NOT EXISTS {extname}"
+        if extversion:
+            sql += f" WITH VERSION {extversion}"
+        if LooseVersion(self.pgversion) > "9.5":
+            sql += " CASCADE"
+        try:
+            with self.cursor(username=self.superuser, dbname=dbname) as cur:
+                cur.execute(sql)
+                # wait until extension is installed
+                timer = Timer(timeout=10, what=f"{extname} is installed")
+                while timer.loop(log=self.log):
+                    cur.execute("SELECT * FROM pg_catalog.pg_extension WHERE extname = %s", (extname, ))
+                    ext = cur.fetchone()
+                    if ext:
+                        self.log.info(
+                            "Installed extension %r to db %r in %r, version: %s", ext["extname"], dbname, self.pgdata,
+                            ext["extversion"]
+                        )
+                        break
+        except psycopg2.OperationalError as err:
+            assert "No such file or directory" in str(err)
+            # extension not available
+            return False
+        else:
+            return True
+
+    def drop_extension(self, *, extname: str, dbname: str):
+        sql = f"DROP EXTENSION IF EXISTS {extname}"
+        if LooseVersion(self.pgversion) > "9.5":
+            sql += " CASCADE"
+        with self.cursor(username=self.superuser, dbname=dbname) as cur:
+            cur.execute(sql)
+
+    def have_aiven_extras(self, *, dbname: str, grantee: str = None):
+        if grantee is None:
+            grantee = self.testuser
+        extname = "aiven_extras"
+        if self.create_extension(extname=extname, dbname=dbname, grantee=grantee):
+            # grant schema usage and select on view
+            with self.cursor(username=self.superuser, dbname=dbname) as cur:
+                cur.execute(f"GRANT USAGE ON SCHEMA {extname} TO {grantee}")
+                cur.execute(f"GRANT SELECT ON aiven_extras.pg_stat_replication TO {grantee}")
+            return True
+        return False
+
+    def drop_aiven_extras(self, *, dbname: str):
+        extname = "aiven_extras"
+        self.drop_extension(extname=extname, dbname=dbname)
+
+    def list_pubs(self, *, dbname: str):
+        """Get all publications created in the database"""
+        with self.cursor(dbname=dbname) as cur:
+            cur.execute("SELECT * FROM pg_catalog.pg_publication")
+            return cur.fetchall()
+
+    def list_slots(self):
+        """Get all replication slots that currently exist on the database cluster"""
+        with self.cursor() as cur:
+            cur.execute("SELECT * FROM pg_catalog.pg_replication_slots")
+            return cur.fetchall()
+
+    def list_subs(self):
+        """Get all existing logical replication subscriptions across all databases of a cluster"""
+        # requires superuser
+        with self.cursor(username=self.superuser) as cur:
+            cur.execute("SELECT * FROM pg_catalog.pg_subscription")
+            return cur.fetchall()
+
+    def list_roles(self):
+        with self.cursor() as cur:
+            cur.execute("SELECT * FROM pg_catalog.pg_roles")
+            return cur.fetchall()
+
+    def add_cleanup(self, cleanup: Callable):
+        self.cleanups.append(cleanup)
 
 
 @contextmanager
